refactor(runtime): extract naming helpers to runtime::naming (Phase 5a)
Begins Phase 5 of the Rust structure refactor roadmap. Per the plan's
explicit guidance, this phase's first step is a standalone naming
extraction that precedes all other runtime concern splits — the
runtime module's internal coupling runs through string-typed image
names, container names, label strings, and family-match helpers, and
centralizing them first means later file extractions (Phase 5b:
repo_cache, image, attach, discovery, cleanup) touch a small stable
interface instead of scattered format strings.

Changes:

- `src/runtime.rs` → `src/runtime/mod.rs` (directory conversion;
  required to host submodules).
- New `src/runtime/naming.rs` (104 LOC) receives:
  - Six Docker label/filter constants:
    `LABEL_MANAGED`, `LABEL_ROLE_AGENT`, `LABEL_ROLE_DIND`,
    `FILTER_MANAGED`, `FILTER_ROLE_AGENT`, `FILTER_ROLE_DIND`.
  - Four naming helpers:
    `image_name`, `dind_certs_volume`, `format_agent_display`,
    and `pub fn matching_family`.
  - Five co-located unit tests (verbatim):
    `dind_certs_volume_derives_from_container_name`,
    `image_name_distinguishes_namespaced_and_flat_classes`,
    `format_agent_display_uses_display_name_for_primary`,
    `format_agent_display_appends_clone_index`,
    `format_agent_display_falls_back_to_container_name`.

Verbatim moves. No function body edits. No renames. No logic changes.
No new types, no newtypes — pure relocation of existing items.

Visibility: the 6 constants and 3 private functions are `pub(super)`
(reachable from `mod.rs` only). `matching_family` stays `pub` and is
re-exported via `pub use self::naming::matching_family;` in `mod.rs`,
so existing `crate::runtime::matching_family` callers in
`src/app/context.rs` and `src/app/mod.rs` resolve unchanged.

Test Preservation: all 418 tests pass, 5 tests moved verbatim with
their original bodies.

Out of scope (deferred to Phase 5b):
- `parse_repo_name`, `git_repo_name`, `claim_container_name`.
- `primary_container_name` (currently in `src/instance.rs`; Phase 7).
- `RUNTIME_OWNED_ENV_VARS` constant (Phase 10 env policy unification).
- Any extraction of repo_cache, image, attach, discovery, cleanup, or
  identity concerns — those belong to subsequent PRs per the plan's
  "do not refactor two large subsystems in one PR" rule.

Verification per COMMITS.md:
- cargo fmt -- --check — pass
- cargo clippy — zero warnings
- cargo nextest run — 418 tests run: 418 passed, 0 skipped (unchanged)
